Updates for phantomjs 1.9.x?
This must have worked for you when you checked it in
but it doesn't seem to work now.

execFile is not sync as it has a callback. That means grunt,
at least current grunt, just exits before the phantomjs continues.

To handle that I needed to add `var done = this.async();` to tell
grunt the rules async and then keep a count and call `done` when
all the images have been processed.

On top of that `child_process.execFile` takes a path to phantomjs
but the code was passing in a path to node. Maybe that's how it
used to work?

Also the current phantomjs provides a path to the exec.

Anyway, with those changes it works for me.